feat(studio): per-simulation LLM temperature control
Add an "LLM Temperature" field (0-2, blank = platform default) to the
edit-simulation Basic Settings, placed after Current State. Renders as a
compact row matching the toggles and clamps input to its bounds; maps
to/from scenarios.metadata.temperature. NUMBER fields now parseFloat so
decimals survive, and support optional min/max bounds.

+++ apps/ally-admin-dashboard/src/utils/common.ts
@@ -237,7 +237,9 @@ export const extractValidData = (
           return [key, isNonEmptyString(value) ? value : null];
 
         case FORM_FIELD_TYPES.NUMBER: //convert string to number and empty val to null
-          return [key, value ? parseInt(value) : null];
+          // parseFloat (not parseInt) so decimal values such as an LLM
+          // temperature of 0.7 survive instead of truncating to 0.
+          return [key, value ? parseFloat(value) : null];
 
         case FORM_FIELD_TYPES.IMAGE_UPLOAD: //image upload if empty returns object,so convert to null
           return [key, value?.length > 0 ? value : null];
